idiomatic errors, fix list subscribers
Instead of returning an int signifying error straight from C, this commit
wraps the return in a function call that will create an error based on the
input string (the function assertGTZero). This also removes errors from
a few functions that actually don't return errors at all in jlog.h.

The LastSubscribers* was meant to *fill in* the char **, not read from.
Rather than creating a char** based off of input slices, I create
a return slice based on the filled char**. Dispose isn't needed
because all it does is free the subscriber char**.