add support for reading and writing message byte slices
This uses reflect.SliceHeader to create []byte or decompose it to void *.
The byte contents do not need to be freed because they are not malloced
in the C file - it uses mmap. This also means the []byte should probably
not have the contents modified.

Includes part of an upcoming ID commit because patch couldn't remove it.