remove GetCheckpoint
The corresponding jlog_get_checkpoint call is in neither the perl nor ruby bindings.
It is in the java binding. This is part of an update where I am removing
the Id struct. The java binding looks like a way to directly wrap the C code,
whereas the ruby and perl bindings abstract away some of the unnecessary low level
details. This one function call would fill in an Id with the current checkpoint
of a subscriber. It'd probably be used in ReadInterval, which will also be removed
shortly. ReadMessage/Read will be the only function calls. Ids will not be manually managed.