move all alter options to inside New
New now includes an Options parameter that sets the file mode, safety and journal
size on creation of the file. This is better than changing it in the middle of
reading/writing (because that is likely not possible). Init is also done inside
of New now because managing New, altering, initing, closing, and newing again
is tedious and best done in one New call. There is no use case to create a file
without initing it.