move all read calls to reader.go
First, this adds a Reader struct in the same manner as the Writer struct.
This also removes all ReadInterval, ReadMessage and ReadCheckpoint calls
and consolidates them into one master Read function that does the same thing
the perl binding's read function does.
Also adds the rewind, checkpoint and autocheckpoint functions that are in
the perl binding.