add concurrency test of multiple readers for a writer
The test fails. The readers each write everything they read to a file. The
md5sums of all the files should be the same. The expected behavior is that
all readers (since they start at the same place) should have the same data
read. This is not the case. When at least one writer has seen and written
everything, the others seem to stop. My guess is that they believe they
have seen the end of the file. They constantly fail on the
'return nil, nil' part of reader.go.
Why is life so hard?